requirements.txt
django==5.1
djangorestframework==3.15.2
pre-commit==3.8.0
black==24.8.0
isort==5.13.2
flake8==7.1.1
django-environ==0.11.2
drf-spectacular==0.27.2
hubspot-api-client==9.0.0
gunicorn==23.0.0